relocate stats middleware function
