+++ cmd/mt-gateway/api.go
@@ -2,10 +2,12 @@ package main
 
 import (
 	"fmt"
+	"github.com/grafana/metrictank/stats"
 	"net/http"
 	"net/http/httputil"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/grafana/metrictank/cmd/mt-gateway/ingest"
 	"github.com/grafana/metrictank/publish"
@@ -78,7 +80,29 @@ func (api Api) Mux() *http.ServeMux {
 
 //Add logging and default orgId middleware to the http handler
 func withMiddleware(svc string, base http.Handler) http.Handler {
-	return defaultOrgIdMiddleware(loggingMiddleware(svc, base))
+	return defaultOrgIdMiddleware(statsMiddleware(loggingMiddleware(svc, base)))
+}
+
+//add request metrics to the given handler
+func statsMiddleware(base http.Handler) http.Handler {
+	stats := requestStats{
+		responseCounts:    make(map[string]map[int]*stats.CounterRate32),
+		latencyHistograms: make(map[string]*stats.LatencyHistogram15s32),
+		sizeMeters:        make(map[string]*stats.Meter32),
+	}
+
+	return http.HandlerFunc(func(w http.ResponseWriter, request *http.Request) {
+		start := time.Now()
+		recorder := responseRecorder{w, -1, 0}
+		base.ServeHTTP(&recorder, request)
+		path := pathSlug(request.URL.Path)
+		stats.PathLatency(path, time.Since(start))
+		stats.PathStatusCount(path, recorder.status)
+		// only record the request size if the request succeeded.
+		if recorder.status < 300 {
+			stats.PathSize(path, recorder.size)
+		}
+	})
 }
 
 //add request logging to the given handler
